Fixed signed/unsigned bug

+++ stratagus/trunk/src/guichan/widgets/window.cpp
@@ -250,7 +250,7 @@ namespace gcn
         
         int textX;
         int textY;
-        textY = (getTitleBarHeight() - getFont()->getHeight()) / 2;
+        textY = ((int)getTitleBarHeight() - getFont()->getHeight()) / 2;
         switch (getAlignment())
         {
           case Graphics::LEFT:
